refactor(dns): split interceptor, simplify nameserver api
reorganised the dns module after the scdynamicstore work had pushed
interceptor.rs past 600 lines with four mixed concerns.

- split interceptor.rs into three submodules. forwarder owns the
  upstream client and per-query handling. upstream owns host dns
  discovery, with a new try_read_scdynamicstore helper isolating the
  macos branch. filter holds the block-list and rebind-protection
  predicates. interceptor.rs is now the thin smoltcp-channel bridge.
- dropped the parse_nameserver free function. fromstr::from_str
  carries the parsing logic directly; the deserialize impl routes
  through s.parse().
- renamed NameserverSpec to Nameserver. the enum self-discriminates
  (addr vs host) and the .resolve() method already signals the
  deferred-resolution semantics, so the spec suffix was redundant.
- extracted a parse_nameservers helper in the node sdk so
  convert_config reads linearly.
- added a comment on the core-foundation 0.9 pin; system-configuration
  0.7 forces that version.
