+++ crates/cli/lib/commands/common.rs
@@ -395,7 +395,10 @@ fn apply_network_opts(
         let dns_nameservers = opts
             .dns_nameserver
             .iter()
-            .map(|s| microsandbox_network::dns::parse_nameserver(s).map_err(anyhow::Error::from))
+            .map(|s| {
+                s.parse::<microsandbox_network::dns::Nameserver>()
+                    .map_err(anyhow::Error::from)
+            })
             .collect::<anyhow::Result<Vec<_>>>()?;
         let dns_query_timeout_ms = opts.dns_query_timeout_ms;
         let network_policy = parse_network_policy(opts.network_policy.as_deref())?;
