+++ crates/network/lib/dns/mod.rs
@@ -6,8 +6,11 @@
 
 pub mod interceptor;
 pub mod parse;
-
 #[cfg(target_os = "macos")]
 pub(crate) mod scdynamicstore;
 
-pub use parse::{NameserverSpec, ParseNameserverError, parse_nameserver};
+mod filter;
+mod forwarder;
+mod upstream;
+
+pub use parse::{Nameserver, ParseNameserverError};
